Add the getAllEntities method to the EntityCache
