+++ engine/src/main/java/org/terasology/entitySystem/entity/EntityCache.java
@@ -144,4 +144,6 @@ public interface EntityCache {
     void destroy(long entityId);
 
     void destroyEntityWithoutEvents(EntityRef entity);
+
+    Iterable<EntityRef> getAllEntities();
 }
